+++ praw/models/util.py
@@ -84,6 +84,7 @@ def stream_generator(
     pause_after: Optional[int] = None,
     skip_existing: bool = False,
     attribute_name: str = "fullname",
+    exclude_before: bool = False,
     **function_kwargs: Any
 ) -> Generator[Any, None, None]:
     """Yield new items from ListingGenerators and ``None`` when paused.
@@ -104,6 +105,9 @@ def stream_generator(
         starting the stream (default: False).
 
     :param attribute_name: The field to use as an id (default: "fullname").
+
+    :param exclude_before: When True does not pass ``params`` to ``functions``
+         (default: False).
 
     Additional keyword arguments will be passed to ``function``.
 
@@ -177,15 +181,9 @@ def stream_generator(
         if before_attribute is None:
             limit -= without_before_counter
             without_before_counter = (without_before_counter + 1) % 30
-        for item in reversed(
-            list(
-                function(
-                    limit=limit,
-                    params={"before": before_attribute},
-                    **function_kwargs
-                )
-            )
-        ):
+        if not exclude_before:
+            function_kwargs["params"] = {"before": before_attribute}
+        for item in reversed(list(function(limit=limit, **function_kwargs))):
             attribute = getattr(item, attribute_name)
             if attribute in seen_attributes:
                 continue